refactoring: Rename uhhm/config.go into config/config.go
I have to admit this one is a personal choice of myself.

It is a little file hardcoding some tool configuration environment
variables and such. Now that it is on the root dir, name it after
config, not the tool's name which could be somewhat misleading.
